Handle heifs complie warning

+++ plugins/common/libs/image/libextplugin/src/heif_impl/heif_parser/heif_parser.cpp
@@ -1252,7 +1252,7 @@ heif_error HeifParser::GetHeifsGroupFrameInfo(uint32_t index, HeifsFrameGroup &f
         return ret;
     }
     sampleNumbers.emplace_back(frameCount + FRAME_INDEX_DELTA);
-    for (int i = 0; i < sampleNumbers.size(); i++) {
+    for (size_t i = 0; i < sampleNumbers.size(); i++) {
         uint32_t keyFrameIndex = sampleNumbers[i] - FRAME_INDEX_DELTA;
         if (index >= keyFrameIndex) {
             beginFrameIndex = keyFrameIndex;
